safestream/requirements.txt
# OCR Engines
paddleocr>=2.7.0
paddlepaddle>=2.5.0
pytesseract>=0.3.10

# Image Processing
opencv-python>=4.8.0
mss>=9.0.0
Pillow>=10.0.0

# OBS Integration
obs-websocket-py>=1.0

# Utilities
pyyaml>=6.0
rapidfuzz>=3.0.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
mypy>=1.5.0
